revisi2

+++ database/migrations/2026_02_07_030502_create_alats_table.php
@@ -11,22 +11,24 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('alat', function (Blueprint $table) {
-            $table->id();
-            $table->string('nama_alat');
-            $table->string('kode_alat')->unique();
-            $table->integer('stok');
-            $table->enum('kondisi', ['baik', 'rusak', 'hilang']); // Assuming enum values
-            $table->string('lokasi_penyimpanan');
-            $table->string('foto_alat')->nullable();
-            $table->text('deskripsi')->nullable();
-            
-            // Foreign Key ke tabel kategori
-            $table->unsignedBigInteger('kategori_id');
-            $table->foreign('kategori_id')->references('id')->on('kategori')->onDelete('cascade');
-            
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('alat')) {
+            Schema::create('alat', function (Blueprint $table) {
+                $table->id();
+                $table->string('nama_alat');
+                $table->string('kode_alat')->unique();
+                $table->integer('stok');
+                $table->enum('kondisi', ['baik', 'rusak', 'hilang']); // Assuming enum values
+                $table->string('lokasi_penyimpanan');
+                $table->string('foto_alat')->nullable();
+                $table->text('deskripsi')->nullable();
+
+                // Foreign Key ke tabel kategori
+                $table->unsignedBigInteger('kategori_id');
+                $table->foreign('kategori_id')->references('id')->on('kategori')->onDelete('cascade');
+
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -34,6 +36,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('alats');
+        Schema::dropIfExists('alat');
     }
 };
