+++ database/migrations/2026_02_07_030516_create_detail_peminjamen_table.php
@@ -11,24 +11,26 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('detail_peminjaman', function (Blueprint $table) {
-            $table->id();
-            
-            // Foreign Key ke tabel peminjaman
-            $table->unsignedBigInteger('peminjaman_id');
-            $table->foreign('peminjaman_id')->references('id')->on('peminjaman')->onDelete('cascade');
-            
-            // Foreign Key ke tabel alat
-            $table->unsignedBigInteger('alat_id');
-            $table->foreign('alat_id')->references('id')->on('alat')->onDelete('cascade');
-            
-            $table->integer('jumlah');
-            $table->integer('jumlah_kembali')->nullable();
-            $table->enum('kondisi_awal', ['baik', 'rusak']);
-            $table->enum('kondisi_kembali', ['baik', 'rusak', 'hilang'])->nullable();
-            
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('detail_peminjaman')) {
+            Schema::create('detail_peminjaman', function (Blueprint $table) {
+                $table->id();
+
+                // Foreign Key ke tabel peminjaman
+                $table->unsignedBigInteger('peminjaman_id');
+                $table->foreign('peminjaman_id')->references('id')->on('peminjaman')->onDelete('cascade');
+
+                // Foreign Key ke tabel alat
+                $table->unsignedBigInteger('alat_id');
+                $table->foreign('alat_id')->references('id')->on('alat')->onDelete('cascade');
+
+                $table->integer('jumlah');
+                $table->integer('jumlah_kembali')->nullable();
+                $table->enum('kondisi_awal', ['baik', 'rusak']);
+                $table->enum('kondisi_kembali', ['baik', 'rusak', 'hilang'])->nullable();
+
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -36,6 +38,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('detail_peminjamen');
+        Schema::dropIfExists('detail_peminjaman');
     }
 };
